# swagger_prod.yaml
swagger: "2.0"
info:
  description: "This is an API based Product Management Application"
  version: "1.0.0"
  title: "Product Management System"
  contact:
    email: "harshitrbsingh@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/"
tags: 
- name: "Product"
  description: "Products that are available in with us"
- name: "Category"
  description: "The Categories into which our products belong"
schemes: 
- "http"
paths:
  /product:
    post:
      tags:
        - "Product"
      summary: "Add a new Product"
      description: "To add a new Product"
      operationId: "addProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in : "body"
        name: "body"
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "Successfully added"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid Input"
          schema:
            $ref: "#/definitions/ApiResponse"
    put:
      tags:
      - "Product"
      summary: "Update an existing Product item using SKU"
      description: "To update an existing Product using its unique SKU value"
      operationId: "updateProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        schema:
          $ref: "#/definitions/Product"
        
      responses:
        200:
          description: "Successfully updated"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/{id}:
    get:
      tags:
      - "Product"
      summary: "Find product by SKU"
      description: "Returns a single product"
      operationId: "getProductByid"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "SKU of the product to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid  SKU supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        404:
          description: "Product not found"
          schema:
            $ref: "#/definitions/ApiResponse"
    delete:
      tags:
      - "Product"
      summary: "Deletes a single product"
      description: ""
      operationId: "deleteProduct"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Product SKU to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successfully deleted"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid SKU supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        404:
          description: "Product not found"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/findByCategory:
    get:
      tags:
      - "Find By"
      summary: "Finds Products by category"
      description: ""
      operationId: "findProductsByCategory"
      produces:
      - "application/json"
      parameters:
      - name: "category"
        in: "query"
        description: "category filter that has to be applied"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
              $ref: "#/definitions/ProdDictResp"
        400:
          description: "Invalid category value"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/findByMode:
    get:
      tags:
      - "Find By"
      summary: "Finds Products by their mode of selling"
      description: ""
      operationId: "findProductsByMode"
      produces:
      - "application/json"
      parameters:
      - name: "mode_of_selling"
        in: "query"
        description: "mode filter that has to be applied"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
              $ref: "#/definitions/ProdDictResp"
        400:
          description: "Invalid mode of selling value"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/findByDemograph:
    get:
      tags:
      - "Find By"
      summary: "Finds Products by Demograph"
      description: ""
      operationId: "findProductsByDemograph"
      produces:
      - "application/json"
      parameters:
      - name: "demograph"
        in: "query"
        description: "Enter the Demograph for which products should be viewed"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
              $ref: "#/definitions/ProdDictResp"
        400:
          description: "Invalid Demograph value"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/findByBrand:
    get:
      tags:
      - "Find By"
      summary: "Finds Products by Brand"
      description: ""
      operationId: "findProductsByBrand"
      produces:
      - "application/json"
      parameters:
      - name: "brand"
        in: "query"
        description: "Brand filter that has to be applied"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
              $ref: "#/definitions/ProdDictResp"
        400:
          description: "Invalid State value"
          schema:
            $ref: "#/definitions/ApiResponse"
  /category:
    post:
        tags:
        - "Category"
        summary: "Add a new category"
        description: ""
        operationId: "addCategory"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in : "body"
          name: "body"
          schema:
            $ref: "#/definitions/Category"
        responses:
          200:
            description: "Successfully added"
            schema:
              $ref: "#/definitions/ApiResponse"
          400:
            description: "Invalid input"
            schema:
              $ref: "#/definitions/ApiResponse"
    delete:
      tags:
      - "Category"
      summary: "Deletes a single Category and all the products with that cid"
      description: ""
      operationId: "deleteCategory"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Category cid to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successfully deleted"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid SKU supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        404:
          description: "Category not found"
          schema:
            $ref: "#/definitions/ApiResponse"

definitions:
  Category:
    type: "object"
    required:
      - id
      - name
      - tax
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
  Product:
    type: "object"
    required:
      - "id"
      - "brand"
      - "category"
      - "sku"
      - "status"
      - "mode_of_selling"
      - "price_inc_tax"
      - "price"
      - "demograph"
      - "metadata"
    properties:
      id:
        type: "string"
        description: "The general name of the product like toothbrush"
      brand:
        type: "string"
        description: "The name of a brand of the product"
      category:
        type: "string"
        description: "The category to which the product belongs"
      sku:
        type: "number"
        description: "The value uniquely identifyig every product"
      status:
        type: "string"
        description: "Whether the product is available or not available"
      mode_of_selling:
        type: "string"
        description: "The mode of selling the product which is whether a product is physical or digital in nature"
      price_inc_tax:
        type: "number"
        description: "The price of a product including 18% price"
      price:
        type: "number"
        description: "The base price of the product without tax"
      demograph:
        type: "string"
        description: "The demograph of the people who use that particular product"
      metadata: 
        type: "string"
        description: "Additional info about the product"
  ApiResponse:
    type: "object"
    properties:
      message:
        type: "string"
  ProdDictResp:
      type: object
      additionalProperties:
        $ref: '#/definitions/Product'
